Fix Add Chapter, change methods, fix aggregation
