+++ server/mongodb/actions/Chapter.ts
@@ -5,9 +5,11 @@ import {
   SearchResponseBody,
   SearchRequestBody,
   IChapterTableEntry,
+  IUser,
+  Role,
 } from "@/common_utils/types";
 import { flatten } from "mongo-dot-notation";
-import { PipelineStage, Promise } from "mongoose";
+import { ObjectId, PipelineStage, Promise } from "mongoose";
 import Chapter from "../models/Chapter";
 import User from "../models/User";
 
@@ -117,7 +119,7 @@ export type ChapterStatistics = {
 export const aggregatePeople = async (
   chapterName: string,
 ): Promise<ChapterStatistics | null> => {
-  const chapterStatistics = await User.aggregate([
+  const chapterStatistics: ChapterStatistics[] = await User.aggregate([
     {
       $facet: {
         activeVolunteers: [
@@ -188,5 +190,73 @@ export const aggregatePeople = async (
     },
   ]);
 
-  return chapterStatistics[0] as ChapterStatistics;
+  await Chapter.findOneAndUpdate(
+    { name: chapterName },
+    {
+      $set: {
+        activeVolunteers: chapterStatistics[0].activeVolunteers,
+        inactiveVolunteers: chapterStatistics[0].inactiveVolunteers,
+        patients: chapterStatistics[0].patients,
+      },
+    },
+    { new: true },
+  );
+
+  return chapterStatistics[0];
+};
+
+export type PostReq = {
+  name: string;
+  chapterPresident: ObjectId;
+  yearFounded: number;
+  country: string;
+  city?: string;
+  state?: string;
+  president: IUser;
+};
+
+export const createChapter = async ({
+  name,
+  chapterPresident,
+  yearFounded,
+  country,
+  state = "",
+  city = "",
+  president,
+}: PostReq): Promise<IChapter> => {
+  await Chapter.create({
+    name,
+    chapterPresident,
+    yearFounded,
+    location: {
+      country,
+      state,
+      city,
+    },
+  });
+
+  await User.findOneAndUpdate<IUser>(
+    { email: president.email },
+    {
+      role: Role.NONPROFIT_CHAPTER_PRESIDENT,
+    },
+  );
+
+  const stats = (await aggregatePeople(name)) as ChapterStatistics;
+
+  const updateFilter = {
+    $set: {
+      activeVolunteers: stats.activeVolunteers,
+      inactiveVolunteers: stats.inactiveVolunteers,
+      patients: stats.patients,
+    },
+  };
+
+  const newStatsChapter: IChapter | null = await Chapter.findOneAndUpdate(
+    { name },
+    updateFilter,
+    { new: true },
+  );
+
+  return newStatsChapter as IChapter;
 };
